+++ block/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @platforma-open/milaboratories.top-antibodies
 
+## 1.1.25
+
+### Patch Changes
+
+- Updated dependencies [456ba67]
+  - @platforma-open/milaboratories.top-antibodies.model@1.7.0
+  - @platforma-open/milaboratories.top-antibodies.ui@1.6.0
+
+## 1.1.24
+
+### Patch Changes
+
+- Updated dependencies [28648b0]
+  - @platforma-open/milaboratories.top-antibodies.ui@1.5.1
+
 ## 1.1.23
 
 ### Patch Changes